% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transtwd97_function.R
\name{distance_TWD97}
\alias{distance_TWD97}
\title{Use TWD97 location data to calculate the distance between two points}
\usage{
distance_TWD97(origin_TWD97_X = NULL, origin_TWD97_Y = NULL,
  observation_TWD97_X = NULL, observation_TWD97_Y = NULL)
}
\arguments{
\item{origin_TWD97_X}{A numeric vector, one of the points' TWD97 X data.}

\item{origin_TWD97_Y}{A numeric vector, one of the points' TWD97 Y data.}

\item{observation_TWD97_X}{A numeric vector, another of the points' TWD97 X data.}

\item{observation_TWD97_X}{A numeric vector, another of the points' TWD97 Y data.}
}
\description{
This function allows you to use TWD97 location data to calculate the distance between two points, create a new vector.
WARNING: ONLY FOR LOCATION OF TAIWAN!!!!!!
}
\examples{
distanceNEW <- distance_TWD97(origin_TWD97_X = 296988.9, origin_TWD97_Y = 2767181, observation_TWD97_X = 196302.7, observation_TWD97_Y = 2828646)
}
\keyword{distance_TWD97}
