% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transtwd97_function.R
\name{distance_WGS84}
\alias{distance_WGS84}
\title{Use WGS84 location data to calculate the distance between two points}
\usage{
distance_WGS84(origin_lat = NULL, origin_lng = NULL,
  obervation_lat = NULL, obervation_lng = NULL)
}
\arguments{
\item{origin_lat}{A numeric vector, one of the points' latitude data.}

\item{origin_lng}{A numeric vector, one of the points' longitude data.}

\item{obervation_lat}{A numeric vector, another of the points' latitude data.}

\item{obervation_lng}{A numeric vector, another of the points' longitude data.}
}
\description{
This function allows you to use WGS84 location data to calculate the distance between two points, create a new vector.
}
\examples{
distanceNEW <- distance_TWD97(origin_lat = 25.01194, origin_lng = 121.4656, obervation_lat = 25.56660, obervation_lng = 120.4656)
}
\keyword{distance_WGS84}
