% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transtwd97_function.R
\name{TWD97toWGS84}
\alias{TWD97toWGS84}
\title{Transforming your TWD97 data to WGS84 data}
\usage{
TWD97toWGS84(inputTWD97Y = NULL, inputTWD97X = NULL, Lng0 = 121)
}
\arguments{
\item{inputTWD97Y}{A numeric vector, TWD97 Y of yours.}

\item{inputTWD97X}{A numeric vector, TWD97 X of yours.}

\item{Lng0}{A numeric vector. If your want to use this function for Penghu, Kinmen and Matsu, you must input new central longitude for 119.}
}
\description{
This function allows you to transform your TWD97 location data to WGS84 data and create a new data frame.
WARNING: ONLY FOR LOCATION OF TAIWAN!!!!!!
}
\examples{
In Taiwan:
transTWD97data <- TWD97toWGS84(inputTWD97Y = 2767181, inputTWD97X = 296988.9)

In Penghu, Kinmen and Matsu:
transTWD97data <- TWD97toWGS84(inputTWD97Y = 2703564, inputTWD97X = 180914.8, Lng0 = 119)
}
\keyword{TWD97toWGS84}
